LangchainThon/requirements.txt
# 추가 설치 추천
python-dotenv 
pydantic 
tiktoken
# P1에서 BM25 앙상블까지 가면
# rank_bm25
# LangSmith 쓰면
# langsmith


# (아래는 기존)
langchain==0.3.27
langchain_community
langchain-openai==0.3.33

langchain-text-splitters
langchain-chroma
langchain-ollama
langchain-experimental

langsmith>=0.1.95

openai

nltk==3.8.1
pypdfium2
pypdf
rapidocr-onnxruntime
sentence-transformers
unstructured
unstructured[pdf]

python-pptx
python-magic-bin==0.4.14

faiss-cpu
grandalf 
streamlit